remove easysetup from main
