+++ src/config.rs
@@ -6,8 +6,6 @@ use std::fs;
 use std::path::{Path, PathBuf};
 use std::str::FromStr;
 
-use crate::provisioning;
-
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Config {
     pub services: Services,
